feat(RunningStatus): modify enum to class-in-enum
